202 & 204 Result codes don't fail run
202 (Accepted) is returned on commit_activity calls, when no data
is available to return. This is a first step on #20, but additional work
needed to rescan to gather the data later.

204 (No Data) is returned on some calls to degenerate repositories. So
far, it's been observed on repos without any commits. The 'no data' was
returned as python 'bytes' which could not be successfully serialized by
tinydb, leading to exceptions. Those are converted to empty lists prior
to further processing.

Also some removeal of code no longer used.