Retry for intermittent error codes
GitHub sometimes returns HTTP codes 403 or 502, which previously failed
the entire run.

Current assumption is that these are indicative of transient conditions
on the GitHub side, so just treat as a retry request. We limit the
number of retries, so this does not become an infinite loop.

Unfortunately, since it is a "production only" error, we'll have to go
live and wait for results.